GbaCell into its own file, also i had broken the rt irq handler.
