+++ src/lib.rs
@@ -4,10 +4,11 @@
 #![feature(isa_attribute)]
 #![feature(naked_functions)]
 
-use core::{
-  cell::UnsafeCell,
-  mem::{align_of, size_of},
-};
+use core::mem::{align_of, size_of};
+
+use gba_cell::GbaCell;
+
+pub mod gba_cell;
 
 #[naked]
 #[no_mangle]
@@ -153,111 +154,22 @@ unsafe extern "C" fn rt0_irq_handler() {
     mov r12, #{mmio_base}
     add r12, r12, #{ime_offset}
     swp r3, r3, [r12]  @IME swap previous
-    mov r0, #{size_of}
-    mov r0, #{align_of}
+    bx lr
     ",
     mmio_base = const 0x0400_0000,
     ime_offset = const 0x208,
     RUST_IRQ_HANDLER = sym RUST_IRQ_HANDLER,
     sys_no_mask = const 0b00011111,
     svc_irq_masked = const 0b10010010,
-    size_of = const size_of::<IrqFn>(),
-    align_of = const align_of::<IrqFn>(),
     options(noreturn)
   )
 }
 
-#[repr(transparent)]
-pub struct GbaCell<T>(UnsafeCell<T>);
-unsafe impl<T> Send for GbaCell<T> {}
-unsafe impl<T> Sync for GbaCell<T> {}
-impl<T> GbaCell<T> {
-  #[inline]
-  pub const fn new(val: T) -> Self {
-    Self(UnsafeCell::new(val))
-  }
-  #[inline]
-  pub fn read(&self) -> T {
-    match (size_of::<T>(), align_of::<T>()) {
-      (4, 4) => unsafe {
-        let val: u32;
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "ldr {r}, [{addr}]",
-          r = out(reg) val,
-          addr = in(reg) p,
-          options(nostack)
-        );
-        core::mem::transmute_copy(&val)
-      },
-      (2, 2) => unsafe {
-        let val: u16;
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "ldrh {r}, [{addr}]",
-          r = out(reg) val,
-          addr = in(reg) p,
-          options(nostack)
-        );
-        core::mem::transmute_copy(&val)
-      },
-      (1, 1) => unsafe {
-        let val: u8;
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "ldrb {r}, [{addr}]",
-          r = out(reg) val,
-          addr = in(reg) p,
-          options(nostack)
-        );
-        core::mem::transmute_copy(&val)
-      },
-      _ => {
-        unimplemented!()
-      }
-    }
-  }
-  #[inline]
-  pub fn write(&self, val: T) {
-    match (size_of::<T>(), align_of::<T>()) {
-      (4, 4) => unsafe {
-        let u: u32 = core::mem::transmute_copy(&val);
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "str {val}, [{addr}]",
-          val = in(reg) u,
-          addr = in(reg) p,
-          options(nostack)
-        )
-      },
-      (2, 2) => unsafe {
-        let u: u16 = core::mem::transmute_copy(&val);
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "strh {val}, [{addr}]",
-          val = in(reg) u,
-          addr = in(reg) p,
-          options(nostack)
-        )
-      },
-      (1, 1) => unsafe {
-        let u: u8 = core::mem::transmute_copy(&val);
-        let p: *mut T = self.0.get();
-        core::arch::asm!(
-          "strb {val}, [{addr}]",
-          val = in(reg) u,
-          addr = in(reg) p,
-          options(nostack)
-        )
-      },
-      _ => {
-        unimplemented!()
-      }
-    }
-  }
-}
-
-#[repr(align(4))]
+/// A function you want called during an interrupt.
+///
+/// This wrapper type is required to make the function pointer be aligned to 4
+/// (normally function pointers are align 1).
+//#[repr(align(4))]
 pub struct IrqFn(pub Option<unsafe extern "C" fn(u16)>);
 
 /// Should store an `unsafe extern "C" fn(u16)`
